Running a seamless console window is not supported due to hooks fails to work properly, mainly because of that the window is owned by csrss.exe process and messagepump works different.
This commit makes it more aware that running cmd.exe though
seamlessshell.exe is not permitted with one exception; when using
the /c <command> argument to start a subprocess.

Fixed a race on windows 2008r2 where the command window still was
hooked and processed that showed a black stale window when application
started using cmd /c <app>. On Windows2003 and earlier this wasn't
a problem because when SetHooks() was reached the console window was
already closed.

git-svn-id: svn://svn.code.sf.net/p/rdesktop/code/seamlessrdp/trunk@1668 423420c4-83ab-492f-b58f-81f9feb106b5